Connect stocks via flows using algebraic notation

The It can be chained: stock1

Usage

lhs %->% rhs

Arguments

lhs

Left-hand side: stock_spec or flow_spec (or NULL for external inflow)

rhs

Right-hand side: flow_spec or stock_spec (or NULL for external outflow)

Value

A partial or complete flow connection

Details

As of version 0.2.0, stocks referenced in flow connections are automatically included in the diagram when finalize() is called, eliminating the need to explicitly add them with fully backward compatible.

Examples

if (FALSE) { # \dontrun{
# Define stocks
S <- stock("S", initial = 1000)
I <- stock("I", initial = 1)

# Define flow
infection <- flow("infection",
  rate = function(inputs, params) params$beta * inputs$S * inputs$I
)

# Connect: S %->% infection %->% I
connection <- S %->% infection %->% I

# Stocks are automatically included when finalized
diagram <- finalize(connection)  # S and I automatically added!

# External flows use NULL
inflow <- NULL %->% some_flow %->% Stock
outflow <- Stock %->% some_flow %->% NULL
} # }